Extract sdl_handle_event
