+++ includes/sdl/sdl_events.h
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   sdl_events.h                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/09/05 15:10:07 by oadhesiv          #+#    #+#             */
+/*   Updated: 2021/09/05 15:10:29 by oadhesiv         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef SDL_EVENTS_H
+# define SDL_EVENTS_H
+
+# include "libft.h"
+# include "sdl_defines.h"
+
+t_byte	sdl_handle_event(t_sdl *sdl);
+
+#endif
