+++ srcs/client/main.c
@@ -6,7 +6,7 @@
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/01 01:06:13 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/09/05 14:45:52 by oadhesiv         ###   ########.fr       */
+/*   Updated: 2021/09/05 15:11:12 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include "init_rtv.h"
 #include "loop_hook.h"
 #include "client/main.h"
-#include "init_sdl.h"
+#include "sdl/init_sdl.h"
 #include "client/network_prepare.h"
 
 void	run_cluster(t_rtv *rtv)
