+++ srcs/loop/loop_hook_redraw.c
@@ -6,12 +6,12 @@
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/11 10:39:04 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/08/28 17:05:26 by oadhesiv         ###   ########.fr       */
+/*   Updated: 2021/09/05 15:11:15 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "loop_hook.h"
-#include "init_sdl.h"
+#include "sdl/init_sdl.h"
 
 static void	*run_parametrized(void *params)
 {
