+++ srcs/main.c
@@ -6,12 +6,12 @@
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/06 15:52:15 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/09/05 14:32:12 by oadhesiv         ###   ########.fr       */
+/*   Updated: 2021/09/05 15:11:10 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "main.h"
-#include "init_sdl.h"
+#include "sdl/init_sdl.h"
 #include "imgui/imgui_init.h"
 
 void	check_defines(void)
